converted putnbr function to receive long so it does not convert uint to int

+++ include/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+        */
+/*   By: juramos <juramos@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 10:35:27 by juramos           #+#    #+#             */
-/*   Updated: 2023/12/07 10:48:47 by juramos          ###   ########.fr       */
+/*   Updated: 2023/12/12 10:54:56 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,5 @@ size_t	ft_write_x(va_list args);
 size_t	ft_write_x_upper(va_list args);
 size_t	ft_write_u(va_list args);
 size_t	ft_write_ptr(va_list args);
-size_t	putnbr_base_with_count_fd(int n, char *base, int fd);
+size_t	putnbr_base_with_count_fd(long n, char *base, int fd);
 #endif
